Move payload handler/sources types to constants
These will be used on plenty of places in the whole payload module so
set them on independent place.
