+++ tests/nosetests/pyanaconda_tests/module_payload_test.py
@@ -31,8 +31,8 @@ from pyanaconda.modules.common.constants.objects import PAYLOAD_DEFAULT, LIVE_OS
 from pyanaconda.modules.payload.base.utils import create_root_dir, write_module_blacklist, \
     get_dir_size
 from pyanaconda.modules.payload.base.initialization import PrepareSystemForInstallationTask
-from pyanaconda.modules.payload.factory import HandlerType, HandlerFactory, SourceType, \
-    SourceFactory
+from pyanaconda.modules.payload.factory import HandlerFactory, SourceFactory
+from pyanaconda.modules.payload.base.constants import HandlerType, SourceType
 from pyanaconda.modules.payload.payload_interface import PayloadInterface
 from pyanaconda.modules.payload.payload import PayloadModule
 from pyanaconda.modules.payload.dnf.dnf import DNFHandlerModule
